requirements/default.txt
amqp==5.0.5
celery==5.0.5
docker~=5.0.0
kubernetes~=12.0.1
python-dotenv==0.16.0
PyYAML~=5.4.1
redis==3.5.3
